src/main/java/courier/Courier.java
package courier;

public class Courier {
    private  String login;
    private  String password;
    private  String firstName;

    public Courier() {
    }

    public Courier(String login, String password, String firstName){
        this.login = login;
        this.password = password;
        this.firstName = firstName;
    }

    static Courier generic() {
        var courier = new Courier("Sanch", "1234", "Lion");
        return courier;
    }

    public String getLogin() {
        return login;
    }

    public void setLogin(String login) {
        this.login = login;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }
}
